refactor: Refactor naming convention
