+++ client/src/components/layout/NavBar/NavBar.component.tsx
