+++ client/src/components/layout/NavBar/index.ts
@@ -1,1 +1,1 @@
-export { NavBar } from "./NavBar";
+export { NavBar } from "./NavBar.component";
