+++ client/src/components/ui/Note/Note.component.tsx
