+++ client/src/components/ui/Note/index.ts
@@ -1,1 +1,1 @@
-export { Note } from "./Note";
+export { Note } from "./Note.component";
