+++ client/src/components/ui/NoteGrid/NoteGrid.component.tsx
