+++ client/src/components/ui/NoteGrid/index.ts
@@ -1,1 +1,1 @@
-export { NoteGrid } from "./NoteGrid";
+export { NoteGrid } from "./NoteGrid.component";
