--- client/src/pages/Home/index.ts
@@ -1,1 +1,0 @@
-export { HomePage } from "./HomePage";
