--- client/src/pages/NewNote/index.ts
@@ -1,1 +1,0 @@
-export { NewNotePage } from "./NewNote";
