--- client/src/pages/Note/index.ts
@@ -1,1 +1,0 @@
-export { NotePage } from "./Note";
