--- client/src/pages/NoteNotFound/index.ts
@@ -1,1 +1,0 @@
-export { NoteNotFoundPage } from "./NoteNotFoundPage";
