--- client/src/pages/Signup/index.ts
@@ -1,1 +1,0 @@
-export { SignupPage } from "./SignupPage";
