+++ client/src/pages/index.ts
@@ -1,7 +1,7 @@
-export { HomePage } from "./Home";
-export { LoginPage } from "./Login";
-export { NewNotePage } from "./NewNote";
-export { NotFoundPage } from "./NotFound";
-export { NotePage } from "./Note";
-export { NoteNotFoundPage } from "./NoteNotFound";
-export { SignupPage } from "./Signup";
+export { HomePage } from "./home";
+export { LoginPage } from "./login";
+export { NewNotePage } from "./new-note";
+export { NotFoundPage } from "./not-found";
+export { NotePage } from "./note";
+export { NoteNotFoundPage } from "./note-not-found";
+export { SignupPage } from "./signup";
