+++ client/src/pages/new-note/NewNote.page.tsx
