+++ client/src/pages/new-note/NewNote.types.ts
