+++ client/src/pages/new-note/index.ts
@@ -1,0 +1,1 @@
+export { NewNotePage } from "./NewNote.page";
