+++ client/src/pages/note-not-found/index.ts
@@ -1,0 +1,1 @@
+export { NoteNotFoundPage } from "./NoteNotFound.page";
