+++ client/src/pages/note/Note.page.tsx
