+++ client/src/pages/note/index.ts
@@ -1,0 +1,1 @@
+export { NotePage } from "./Note.page";
