+++ client/src/pages/signup/Signup.types.ts
