+++ client/src/pages/signup/index.ts
@@ -1,0 +1,1 @@
+export { SignupPage } from "./Signup.page";
